# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusterhubGo",
        "description": "This api is for working with musterhabGo",
        "contact": {
            "email": "artem.yablochnyi@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Business service"
                ],
                "summary": "Services list",
                "description": "Returns services list",
                "operationId": "services",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items to return",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServiceCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/business-profiles/{business_profile}/services": {
            "get": {
                "tags": [
                    "Business profile"
                ],
                "summary": "Services list for given business profile",
                "description": "Returns services list for given business profile",
                "operationId": "business-services",
                "parameters": [
                    {
                        "name": "business_profile",
                        "in": "path",
                        "description": "Business profile id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items to return",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServiceCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Business profile"
                ],
                "summary": "Creates new service for given business profile",
                "description": "Returns created service data",
                "operationId": "business-services-create",
                "parameters": [
                    {
                        "name": "business_profile",
                        "in": "path",
                        "description": "Business profile id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusinessServiceCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServiceResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Business service"
                ],
                "summary": "Get service information",
                "description": "Returns service data by id",
                "operationId": "service-id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServiceResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Business service"
                ],
                "summary": "Updates service by id",
                "description": "Returns updated service",
                "operationId": "service-update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusinessServiceUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServiceResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Business service"
                ],
                "summary": "Deletes service by id",
                "description": "Deletes service by id",
                "operationId": "service-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BusinessService": {
                "title": "BusinessService",
                "description": "BusinessService model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "business_id": {
                        "title": "business_id",
                        "description": "Related business profile id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Service title",
                        "type": "string"
                    },
                    "price": {
                        "$ref": "#/components/schemas/Price"
                    },
                    "description": {
                        "title": "description",
                        "description": "Service description",
                        "type": "string",
                        "nullable": true
                    },
                    "duration": {
                        "title": "duration",
                        "description": "Duration in minutes",
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "is_field": {
                        "title": "is_field",
                        "description": "Indicates whether the service can be provided in the fields",
                        "type": "boolean"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "Owner id",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Date and time of creation",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-05-28T13:51:36.000000Z"
                    },
                    "updated_at": {
                        "title": "description",
                        "description": "Date and time of updation",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-05-28T13:51:36.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessService"
                }
            },
...
            "Currency": {
                "title": "Currency",
                "description": "Currency model. Implements country currency codes according to ISO 4217 standard",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "alpha_code": {
                        "title": "Alpha code",
                        "description": "Currency alphabetical code",
                        "type": "string",
                        "example": "RUB"
                    },
                    "numeric_code": {
                        "title": "Numeric code",
                        "description": "Currency numeric code",
                        "type": "string",
                        "example": "643"
                    },
                    "fraction": {
                        "title": "Fraction",
                        "description": "Number of fractional units (e.g. in dollar 100 cents)",
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0,
                        "multipleOf": 10,
                        "example": 100
                    },
                    "sign": {
                        "title": "Sign",
                        "description": "Currency sign",
                        "type": "string",
                        "example": "₽",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Currency"
                }
            },
...
            "BusinessServiceCreateRequest": {
                "title": "BusinessServiceCreateRequest",
                "description": "Creates service",
                "required": [
                    "title",
                    "price",
                    "currency_id",
                    "duration",
                    "is_field",
                    "user_id"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Service title",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "description": "Price",
                        "type": "number",
                        "minimum": 0,
                        "example": 12.3
                    },
                    "currency_id": {
                        "title": "currency_id",
                        "description": "Currency id",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "description": {
                        "title": "description",
                        "description": "Service description",
                        "type": "string",
                        "nullable": true
                    },
                    "duration": {
                        "title": "duration",
                        "description": "Duration in minutes",
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1,
                        "example": 45
                    },
                    "is_field": {
                        "title": "is_field",
                        "description": "Indicates whether the service can be provided in the fields",
                        "type": "boolean"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "Owner id",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "BusinessServiceUpdateRequest": {
                "title": "BusinessServiceUpdateRequest",
                "description": "Updates service",
                "required": [],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Service title",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "description": "Price",
                        "type": "integer",
                        "minimum": 0,
                        "example": 12.99
                    },
                    "currency_id": {
                        "title": "currency_id",
                        "description": "Currency id",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "description": {
                        "title": "description",
                        "description": "Service description",
                        "type": "string",
                        "nullable": true
                    },
                    "duration": {
                        "title": "duration",
                        "description": "Duration in minutes",
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1,
                        "example": 45
                    },
                    "is_field": {
                        "title": "is_field",
                        "description": "Indicates whether the service can be provided in the fields",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "BusinessServiceCollection": {
                "title": "BusinessServiceCollection",
                "description": "BusinessService collection",
                "properties": {
                    "count": {
                        "title": "Count",
                        "description": "Number of items in response",
                        "type": "integer",
                        "example": 1
                    },
                    "items": {
                        "title": "Items",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BusinessService"
                        }
                    },
                    "total": {
                        "title": "Total",
                        "description": "Number of items in database",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessServiceCollection"
                }
            },
            "BusinessServiceResource": {
                "title": "BusinessServiceResource",
                "description": "BusinessService resource",
                "properties": {
                    "item": {
                        "$ref": "#/components/schemas/BusinessService"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessServiceResource"
                }
            },
...
            "UnprocessableContentResponse": {
                "title": "UnprocessableContentResponse",
                "description": "Response when validation failed",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Error message",
                        "type": "string"
                    },
                    "errors": {
                        "$ref": "#/components/schemas/UnprocessableContentResponseError"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UnprocessableContentResponse"
                }
            },
            "UnprocessableContentResponseError": {
                "title": "Error",
                "description": "Error",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Faild field name",
                        "type": "array",
                        "items": {
                            "description": "Error message",
                            "type": "string"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Error"
                }
            },
            "Price": {
                "title": "Price",
                "description": "Price",
                "properties": {
                    "amount": {
                        "title": "amount",
                        "description": "Amount",
                        "type": "number",
                        "example": 129.99
                    },
                    "currency": {
                        "$ref": "#/components/schemas/Currency"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Price"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}